requirements.txt
tqdm
py7zr
patool
pyyaml